remove Utils class and its subpackage
All Nethods in Utils.java directly use the XMLRPC and are not equivalently translatable to
the REST-API as we would have to handle every XMLRPC endpoint with special code on the
client side. The remaining XMLRPC methods have been moved to the DocpoolBaseService class.
As future REST-API methods will depend on the service object anyway, there is currently no need
for an external utils class.