remove exception subpackage
The only exception defied there is never thrown. Given that we need a drop-in replacement
we cannot throw any exceptions from the methods published so far let alone custom
new exceptions. Futhermore, we do not expect so many new exceptions such that they should
be in their own package (which hinders code sharing).